Added option to make library
Now one can generate a library `libcoffe.a`,
which can also be installed (along with the binary `coffe`)
on the system. A simple example:

#include <coffe/common.h>
#include <coffe/parser.h>

int main(void){
    struct coffe_parameters_t par;
    coffe_parse_default_parameters(&par);
    coffe_parameters_free(&par);

    return 0;
}

The above should be compiled with the required libraries,
which are as follows on GCC 8:
-lconfig -lm -lgsl -lcoffe
The other optional libraries needed, depending on the usage, are:
-lcuba -lclass -lfftw3
Additionally, `-fopenmp` may be needed when compiling the program
if the library was compiled with it as well.