Fixed segfault if one links CLASS in the lib
If one makes a test program that uses the COFFE lib,
and the lib has been compiled with CLASS, it would
randomly segfault because the macro `HAVE_CLASS`
wasn't defined in the test program, but WAS defined
at the time when the lib was built, which would cause
some elements of the `coffe_parameters_t` struct,
which is declared in `common.h`, to not actually exist
when running the test program, causing a very strange
segfault.
The easiest fix is to just not use that macro in the
header and the default parameters parser, which is
exactly what was done in this commit.