Growth function is now normalized
`D1`, representing the growth function, is now such that D1(z = 0) = 1.
This is a breaking change with respect to the previous version of COFFE,
and as a result, one of the tests had to be modified.
This was done to facilitate compatibility with CLASS, which already has
a normalized growth function.