Revamped flatsky integral
It can be obtained from the FFTlog transform
since the spherical Bessel function is just
a special case of the Bessel function,
but in order for it to work, the parameter l
must be a double instead of an int.
The integrals have also been independently verified
with a Python code.

Also removed redundant flatsky function from `twofast` module.