Fixed parsing of `d2`. Unified parsers.
There was a typo in `parser.c` which caused the
contribution of `d2` to not show up in the results.
Additionally, the settings between the default parser,
the `settings.cfg` file, and the `coffe.py` wrapper
have now (hopefully) been set to the same values,
so running via Python and the using the default `settings.cfg`
file should yield more or less the same output.