Big changes in the the way integrals are computed
This is a pretty big commit (and hopefully everything works!).
I've replaced the inflexible `coffe_integrals_t` type
with a `coffe_integral_array_t` type, which just contains
a pointer to a `coffe_integral_t` type, containing the results
of the integral (the spline), any possible renormalization terms,
as well as the indices n, l. Additionally, since in flat-sky we need
ordinary, and NOT spherical Bessel functions, I've added two flags,
denoting whether or not the values should be divided by a factor of 2.
To make things easier, I've made a helper function,
`coffe_find_integral`, which just returns a pointer to the
`coffe_integral_t` type, and its members, `result`, renormalization`, or
`renormalization_zero_separation`, can be directly passed to the various
interpolation functions (if nothing is found, it returns NULL, which may
cause a segfault if the user is not careful).
Finally, I've added helper functions for handling arrays, so I don't
need to spam for loops all over the place, hopefully making it easier to
tell what's happening.
I probably still need to format the code so not everything is in a
single line.