Added windowed covariance
Implementation follows eq. 18 from arXiv:1509.04293. The largest codebase
change was for the Poisson x Poisson term, which instead of a Kronecker delta
function now requires the evaluation of an integral over the two windowed
spherical Bessel functions. This seems to have some impact on the runtime when
using the standard GSL integrator, but not so much with the double exponential
quadrature.