Removed Cuba parallellization in interface
Leaving in the paralellization causes the code to hand when running the OpenMP
loops, so that one has to be disabled then.  Running some tests, it seems the
parallel multidimensional sampling is actually slower than just using OpenMP,
so it'd be ideal to disable Cuba paralellization, which is what this commit
achieves by just setting the `CUBACORES` environmental variable to 0, which
works.
Strangely enough, if I directly call `cubacores` from the C code, and set
`CUBACORES` to 0, which should have the same effect, it crashes the Cython code
with a very strange error:

python3.7: ./src/common/Fork.c:100: cubafork: Assertion `socketpair(AF_LOCAL,
SOCK_STREAM, 0, fd) != -1 && (pid = fork()) != -1' failed.

I've tried experimenting with the Cython `nogil` option, but to no avail, so
this is more of a hack than anything (that works!).